a11y: correctly announce notifications to screen readers
server-rendered notifications where not always announced correctly (for
example because we autofocus inputs).

make sure they are announced at load so that screen reader users
correctly understand when there are warnings or errors after their
actions.
