--- assets/js/modules/sr-external-links.js
@@ -1,8 +1,0 @@
-export const explainExternalLinks = () => {
-  document.querySelectorAll('a[target="_blank"]').forEach((link) => {
-    const ariaLabel = link.getAttribute("aria-label");
-    if (!ariaLabel) {
-      link.setAttribute("aria-label", link.textContent + " (nouvelle fenêtre)");
-    }
-  });
-};
