+++ src/views/partials/notifications.ejs
@@ -1,4 +1,5 @@
 <% if (notifications && notifications.length > 0) { %>
+  <div id="notifications">
     <% if (!locals.noWrapperDiv) { %>
     <div class="fr-input-group">
         <% } %>
@@ -13,4 +14,5 @@
         <% if (!locals.noWrapperDiv) { %>
     </div>
     <% } %>
+  </div>
 <% } %>
