+++ src/views/user/sign-in-with-passkey.ejs
@@ -1,7 +1,7 @@
 <div>
     <%- include('../partials/notifications.ejs', {notifications: notifications, noWrapperDiv: true}) %>
-    <div class="fr-alert fr-alert--error fr-mb-2w" id="webauthn-alert-error" hidden></div>
-    <div class="fr-alert fr-alert--success fr-mb-2w" id="webauthn-alert-success" hidden></div>
+    <div class="fr-alert fr-alert--error fr-mb-2w sr-announce" role="alert" id="webauthn-alert-error" hidden></div>
+    <div class="fr-alert fr-alert--success fr-mb-2w sr-announce" role="status" id="webauthn-alert-success" hidden></div>
     <h1 class="fr-h3">
         Se connecter avec une clé d’accès
     </h1>
